Updated float32 to float64

+++ workspace/src/par_pkg/par_pkg/board_transformer_node.py
@@ -8,7 +8,7 @@ from geometry_msgs.msg import PoseStamped
 from par_interfaces.msg import IVector2
 from par_interfaces.srv import BoardToWorld, WorldToBoard
 import math
-from std_msgs.msg import Bool, Float32MultiArray, Float32
+from std_msgs.msg import Bool, Float32MultiArray, Float64
 from rclpy.qos import ReliabilityPolicy, QoSProfile
 
 ## This node updates the board transform,
@@ -48,7 +48,7 @@ class BoardTransformerNode(Node):
 
         self.__objects_subscription = self.create_subscription(Float32MultiArray, 'objects', self.__object_callback,  qos_profile=QoSProfile(depth=10, reliability=ReliabilityPolicy.RELIABLE))
 
-        self.__table_height_publisher = self.create_publisher(Float32, '/par/table_height',qos_profile=QoSProfile(depth=10, reliability=ReliabilityPolicy.RELIABLE))
+        self.__table_height_publisher = self.create_publisher(Float64, '/par/table_height',qos_profile=QoSProfile(depth=10, reliability=ReliabilityPolicy.RELIABLE))
 
         self.__board_found_publisher = self.create_publisher(Bool, "/par/board_found", qos_profile=QoSProfile(depth=10, reliability=ReliabilityPolicy.RELIABLE))
 
@@ -133,7 +133,7 @@ class BoardTransformerNode(Node):
         
         self.__broadcaster.sendTransform(self.__board_transform)
 
-        table_height_msg = Float32()
+        table_height_msg = Float64()
         table_height_msg.data = self.__board_transform.transform.translation.z
         self.__table_height_publisher.publish(table_height_msg.data)
